Audit.AzureCosmos unit tests
